+++ policyengine_uk/variables/household/income/household_benefits.py
@@ -49,6 +49,8 @@ class household_benefits(Variable):
         "dfe_education_spending",
         "dft_subsidy_spending",
         "healthy_start_vouchers",
+        "two_child_limit_payment",
+        "scottish_child_payment",
     ]
 
     def formula(household, period, parameters):