+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.68.0] - 2026-01-13 15:46:23
+
+### Added
+
+- 2025-26 Scottish income tax rates.
+
+## [2.67.0] - 2026-01-13 14:23:42
+
+### Added
+
+- Scottish Child Payment - a benefit provided by Social Security Scotland for eligible children under 16 in low-income families receiving qualifying benefits (UC, tax credits, etc.).
+
+## [2.66.0] - 2026-01-12 16:12:34
+
+### Fixed
+
+- LHA freeze parameter.
+
 ## [2.65.9] - 2025-12-09 16:16:37
 
 ### Fixed
@@ -2487,6 +2505,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[2.68.0]: https://github.com/PolicyEngine/openfisca-uk/compare/2.67.0...2.68.0
+[2.67.0]: https://github.com/PolicyEngine/openfisca-uk/compare/2.66.0...2.67.0
+[2.66.0]: https://github.com/PolicyEngine/openfisca-uk/compare/2.65.9...2.66.0
 [2.65.9]: https://github.com/PolicyEngine/openfisca-uk/compare/2.65.8...2.65.9
 [2.65.8]: https://github.com/PolicyEngine/openfisca-uk/compare/2.65.7...2.65.8
 [2.65.7]: https://github.com/PolicyEngine/openfisca-uk/compare/2.65.6...2.65.7